+++ lib/extensions/base.py
@@ -3,13 +3,9 @@
 # Copyright (c) 2010 Authors
 # Licensed under the GNU GPL version 3.0 or later.  See the file LICENSE for details.
 
-import json
 import os
-import re
-from collections.abc import MutableMapping
 
 import inkex
-from lxml import etree
 from lxml.etree import Comment
 from stringcase import snakecase
 
@@ -18,95 +14,21 @@ from ..elements import EmbroideryElement, nodes_to_elements
 from ..elements.clone import is_clone
 from ..i18n import _
 from ..marker import has_marker
+from ..metadata import InkStitchMetadata
 from ..svg import generate_unique_id
 from ..svg.tags import (CONNECTOR_TYPE, EMBROIDERABLE_TAGS, INKSCAPE_GROUPMODE,
                         NOT_EMBROIDERABLE_TAGS, SVG_CLIPPATH_TAG, SVG_DEFS_TAG,
                         SVG_GROUP_TAG, SVG_MASK_TAG)
-from ..update import update_legacy_params
-from ..utils.settings import DEFAULT_METADATA, global_settings
-
-INKSTITCH_SVG_VERSION = 1
-
-
-def strip_namespace(tag):
-    """Remove xml namespace from a tag name.
-
-    >>> {http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd}namedview
-    <<< namedview
-    """
-
-    match = re.match(r'^\{[^}]+\}(.+)$', tag)
-
-    if match:
-        return match.group(1)
-    else:
-        return tag
-
-
-class InkStitchMetadata(MutableMapping):
-    """Helper class to get and set inkstitch-specific metadata attributes.
-
-    Operates on a document and acts like a dict.  Setting an item adds or
-    updates a metadata element in the document.  Getting an item retrieves
-    a metadata element's text contents or None if an element by that name
-    doesn't exist.
-    """
-
-    def __init__(self, document):
-        super().__init__()
-        self.document = document
-        self.metadata = document.metadata
-
-        for setting in DEFAULT_METADATA:
-            if self[setting] is None:
-                self[setting] = global_settings[f'default_{setting}']
-
-    # Because this class inherints from MutableMapping, all we have to do is
-    # implement these five methods and we get a full dict-like interface.
-    def __setitem__(self, name, value):
-        item = self._find_item(name)
-        item.text = json.dumps(value)
-
-    def _find_item(self, name, create=True):
-        tag = inkex.addNS(name, "inkstitch")
-        item = self.metadata.find(tag)
-        if item is None and create:
-            item = etree.SubElement(self.metadata, tag)
-
-        return item
-
-    def __getitem__(self, name):
-        item = self._find_item(name)
-
-        try:
-            return json.loads(item.text)
-        except (ValueError, TypeError):
-            return None
-
-    def __delitem__(self, name):
-        item = self._find_item(name, create=False)
-
-        if item is not None:
-            self.metadata.remove(item)
-
-    def __iter__(self):
-        for child in self.metadata:
-            if child.prefix == "inkstitch":
-                yield strip_namespace(child.tag)
-
-    def __len__(self):
-        i = 0
-        for i, item in enumerate(self):
-            pass
-
-        return i + 1
-
-    def __json__(self):
-        return dict(self)
+from ..update import update_inkstitch_document
 
 
 class InkstitchExtension(inkex.EffectExtension):
     """Base class for Inkstitch extensions.  Not intended for direct use."""
+
+    def load(self, *args, **kwargs):
+        document = super().load(*args, **kwargs)
+        update_inkstitch_document(document)
+        return document
 
     @classmethod
     def name(cls):
@@ -189,12 +111,7 @@ class InkstitchExtension(inkex.EffectExtension):
     def get_nodes(self, troubleshoot=False):
         # Postorder traversal of selected nodes and their descendants.
         # Returns all nodes if there is no selection.
-        descendants = self.descendants(self.document.getroot(), troubleshoot=troubleshoot)
-
-        # update nodes as necessary
-        self._update_inkstitch_document(descendants)
-
-        return descendants
+        return self.descendants(self.document.getroot(), troubleshoot=troubleshoot)
 
     def get_elements(self, troubleshoot=False):
         self.elements = nodes_to_elements(self.get_nodes(troubleshoot))
@@ -227,45 +144,3 @@ class InkstitchExtension(inkex.EffectExtension):
     def uniqueId(self, prefix, make_new_id=True):
         """Override inkex.Effect.uniqueId with a nicer naming scheme."""
         return generate_unique_id(self.document, prefix)
-
-    def _update_inkstitch_document(self, nodes):
-        # get the inkstitch svg version from the document
-        svg = self.document.getroot()
-        search_string = "//*[local-name()='inkstitch_svg_version']//text()"
-        file_version = svg.findone(search_string)
-        if file_version is None:
-            file_version = 0
-        else:
-            file_version = int(file_version)
-
-        if file_version == INKSTITCH_SVG_VERSION:
-            return
-
-        if file_version > INKSTITCH_SVG_VERSION:
-            inkex.errormsg(_("This document was created with a newer Version of Ink/Stitch. "
-                             "It is possible that not everything works as expected.\n\n"
-                             "Please update your Ink/Stitch version: https://inkstitch.org/docs/install/"))
-            # they may not want to be bothered with this info everytime they call an inkstitch extension
-            # let's udowngrade the file version number
-            self._update_inkstitch_svg_version()
-        else:
-            # this document is either a new document or it is outdated
-            # if we cannot find any inkstitch attribute in the document, we assume that this is a new document which doesn't need to be updated
-            search_string = "//*[namespace-uri()='http://inkstitch.org/namespace' or " \
-                            "@*[namespace-uri()='http://inkstitch.org/namespace'] or " \
-                            "@*[starts-with(name(), 'embroider_')]]"
-            inkstitch_element = svg.findone(search_string)
-            if inkstitch_element is None:
-                self._update_inkstitch_svg_version()
-                return
-
-            # update elements
-            elements = nodes_to_elements(nodes)
-            for element in elements:
-                update_legacy_params(element, file_version, INKSTITCH_SVG_VERSION)
-            self._set_inkstitch_svg_version()
-
-    def _update_inkstitch_svg_version(self):
-        # set inkstitch svg version
-        metadata = self.get_inkstitch_metadata()
-        metadata['inkstitch_svg_version'] = INKSTITCH_SVG_VERSION
